refectored to types
